Prepare for Bintray/Maven Central release
Describe important changes and explain the version bump in a changelog.

Add artifact tasks for source and Javadoc jars; fix javadoc error in
StringUtils.java.

Add tasks for Bintray and Maven Central publishing. Retrieve passwords and
usernames from files outside of the project (NOT TO BE COMMITTED!!!). Sign the
distribution, as required by Maven Central.