Put models in separate jar; publish all jars
Clean up the jar generation (set `classifier` instead of `basename` for
`fatJar`) and set all of them for publication. Add a new jar specifically for
the models, and remove the models from the library jar. Keep using them as
resources in idea (don't know how to set that generally for any IDE in use).